# server/client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "linkedin-challenge-server",
    "description": "",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "tags": [
          "signup"
        ],
        "operationId": "signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Organization created; session cookie set",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Password too short",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/orgs/{slug}/admin/invites": {
      "post": {
        "tags": [
          "invites"
        ],
        "operationId": "createInvites",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Codes generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateInvitesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in as an admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "A JSON error body. Every API route returns this shape on failure, so the generated client\nand the extension can rely on `{ \"error\": \"...\" }` regardless of status.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "CreateInvitesRequest": {
        "type": "object",
        "properties": {
          "count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "role": {
            "type": [
              "string",
              "null"
            ],
            "description": "`\"participant\"` (default) or `\"admin\"`."
          }
        }
      },
      "CreateInvitesResponse": {
        "type": "object",
        "required": [
          "codes"
        ],
        "properties": {
          "codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SignupRequest": {
        "type": "object",
        "required": [
          "orgName",
          "name",
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "orgName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "SignupResponse": {
        "type": "object",
        "required": [
          "ok",
          "orgSlug"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "orgSlug": {
            "type": "string"
          }
        }
      },
...
    }
  }
}